backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
PyJWT
flask-swagger-ui
